requirements.txt
aiohttp==3.7.2
altgraph==0.17
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cloudscraper==1.2.48
fake-headers==1.0.2
furl==2.1.0
html5lib==1.1
idna==2.10
macholib==1.14
multidict==5.0.0
orderedmultidict==1.0.1
Pillow==8.0.1
PixivPy==3.5.10
pygelbooru==0.3.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
pysaucenao==1.4.0
requests==2.24.0
requests-toolbelt==0.9.1
saucenao-api==2.3.0
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.11
webencodings==0.5.1
xmltodict==0.12.0
yarl==1.6.2
